cleanup zarr reader to io_utils, add chunking strategy
